Shortcut helper compilation when Cabal version is already available
This is mostly an optimization for Nix which already sets up the environment
correctly, so we should reward its users :)
